fix: avoid `-` in theme name
In the new way that themes are generated by `estilo`, the `-` in the name breaks the syntax file.

Since `_` is considered to be part of a word in VimScript, it is safer to use for the name of the theme.

BREAKING CHANGE: The name of the theme has changed to `night_owl` rather than `night-owl` or `nightowl` (lightline)